refactor(circuit): split sympy helpers out of circuit_utils [Clean Code: SRP]
Move to a new `scqubits/core/sympy_helpers.py`:

- `_generate_symbols_list` (private)
- `is_potential_term`, `round_symbolic_expr`, `keep_terms_for_subsystem`
  (public)

Internal callers (`circuit_routines.py`, `circuit_sym_methods.py`,
`symbolic_circuit.py`, `symbolic_circuit_graph.py`,
`tests/test_circuit_utils.py`) now import from `sympy_helpers`
directly. Public deprecation shims for the three public symbols stay
in `circuit_utils.py` (downstream user code may import them from the
old location).

`keep_terms_for_subsystem` has zero internal callers in the current
tree, but it's public — keep the shim.
