+++ scqubits/core/symbolic_circuit_graph.py
@@ -31,10 +31,8 @@ from scqubits.core.circuit_input import (
     remove_comments,
     strip_empty_lines,
 )
-from scqubits.core.circuit_utils import (
-    get_trailing_number,
-    round_symbolic_expr,
-)
+from scqubits.core.circuit_utils import get_trailing_number
+from scqubits.core.sympy_helpers import round_symbolic_expr
 from scqubits.utils.misc import (
     flatten_list_recursive,
     unique_elements_in_list,
